requirements.txt
mkdocs
mkdocs-material
mkdocstrings[python]
pyyaml
mkdocs-gen-files
